Remove unsearchable properties from kw.properties
